requirements.txt
requests>=2.31.0
aiohttp>=3.8.0
pandas>=2.0.0
numpy>=1.24.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
google-generativeai>=0.3.0
mp-api>=0.33.0
python-dotenv>=1.0.0
 